Scanner class are feractored to one class per file
