+++ container_pipeline/workers/scanners/base.py
